Доконфигурировал логгирование и ошибки

+++ src/BeerEconomy.DataStorageService/Database/Repositories/Impl/BeerRepository.cs
@@ -1,3 +1,6 @@
+using System.Text;
+using BeerEconomy.Common.Helpers.Exceptions;
+using BeerEconomy.Common.Helpers.Logging;
 using BeerEconomy.DataStorageService.Database.Entities;
 using Microsoft.EntityFrameworkCore;
 
@@ -8,6 +11,8 @@ namespace BeerEconomy.DataStorageService.Database.Repositories.Impl;
 /// </summary>  
 internal sealed class BeerRepository(DataContext dataContext) : IRepository<BeerEntity>
 {
+    private static readonly ILogger<BeerRepository> Logger = StaticLogger.CreateLogger<BeerRepository>();
+    
     /// <inheritdoc />
     public IQueryable<BeerEntity> CreateQuery()
     {
@@ -19,18 +24,53 @@ internal sealed class BeerRepository(DataContext dataContext) : IRepository<Beer
     /// <inheritdoc />
     public async Task<BeerEntity> GetAsync(int id, CancellationToken cancellationToken)
     {
-        return await CreateQuery().FirstAsync(b => b.Id == id, cancellationToken);
+        var entity = await CreateQuery().FirstOrDefaultAsync(b => b.Id == id, cancellationToken);
+        if (entity == null)
+        {
+            throw new InternalException(ErrorCode.NOT_FOUND, $"Не найдено пиво с id #{id}");
+        }
+
+        return entity;
     }
 
     /// <inheritdoc />
     public async Task<BeerEntity> UpdateAsync(int id, BeerEntity entity, CancellationToken cancellationToken)
     {
         var existingEntity = await GetAsync(id, cancellationToken);
-        existingEntity.Description = entity.Description;
-        existingEntity.ImageUrl = entity.ImageUrl;
-        existingEntity.Name = entity.Name;
+        
+        var log = new LogMessage($"Обновлена сущность пиво #{id}.");
+        var updated = false;
+        
+        if (existingEntity.Name != entity.Name)
+        {
+            log = log.PropertyChanged(nameof(existingEntity.Name), existingEntity.Name, entity.Name);
+            existingEntity.Name = entity.Name;
+            updated = true;
+        }
+        
+        if (existingEntity.Description != entity.Description)
+        {
+            log = log.PropertyChanged(nameof(existingEntity.Description), existingEntity.Description, entity.Description);
+            existingEntity.Description = entity.Description;
+            updated = true;
+        }
+        
+        if (existingEntity.ImageUrl != entity.ImageUrl)
+        {
+            log = log.PropertyChanged(nameof(existingEntity.ImageUrl), existingEntity.ImageUrl, entity.ImageUrl);
+            existingEntity.ImageUrl = entity.ImageUrl;
+            updated = true;
+        }
+
+        if (!updated)
+        {
+            return existingEntity;
+        }
+
         await dataContext.SaveChangesAsync(cancellationToken);
 
+        Logger.LogInformation(log);
+        
         return existingEntity;
     }
 
@@ -39,15 +79,30 @@ internal sealed class BeerRepository(DataContext dataContext) : IRepository<Beer
     {
         await dataContext.AddAsync(entity, cancellationToken);
         await dataContext.SaveChangesAsync(cancellationToken);
+        
+        var log = new LogMessage($"Создана сущность пиво #{entity.Id}.");
+        log = log.Property(nameof(entity.Name), entity.Name)
+            .Property(nameof(entity.Description), entity.Description)
+            .Property(nameof(entity.ImageUrl), entity.ImageUrl);
+        Logger.LogInformation(log);
+        
         return entity;
     }
 
     /// <inheritdoc />
     public async Task DeleteAsync(int id, CancellationToken cancellationToken)
     {
-        await dataContext.Set<BeerEntity>().Where(b => b.Id == id).ExecuteDeleteAsync(cancellationToken);
+        var beerCount = await dataContext.Set<BeerEntity>().Where(b => b.Id == id).ExecuteDeleteAsync(cancellationToken);
+        if (beerCount == 0)
+        {
+            throw new InternalException(ErrorCode.NOT_FOUND, $"Не найдено пиво с id #{id}");
+        }
+        
         await dataContext.Set<SourceEntity>().Where(b => b.BeerId == id).ExecuteDeleteAsync(cancellationToken);
         await dataContext.Set<PriceEntity>().Where(b => b.BeerId == id).ExecuteDeleteAsync(cancellationToken);
         await dataContext.SaveChangesAsync(cancellationToken);
+        
+        var log = new LogMessage($"Удалена сущность пиво #{id} каскадом.");
+        Logger.LogInformation(log);
     }
 }
